+++ MWSE/WindowsUtil.cpp
@@ -14,7 +14,7 @@ namespace mwse::lua {
 		int64_t elapsed = endingTime.QuadPart - startTimestamp;
 		// Conversion to microseconds, we do this first to avoid loss-of-precision during the next division
 		elapsed *= 1000000ll;
-		double t = elapsed / performanceFrequency;
+		double t = static_cast<double>(elapsed / performanceFrequency);
 		t /= 1000000;
 		return t;
 	}
